Create selectionSort.cpp
Main idea of selection sort is that we maintain a variable minIndex, which stores the index of the minimum number in the array with reference to the current index. Inititally minIndex = i, but if found another number which is smallest as we traverse, minIndex is updated and the current  is swapped with number with min index.